refactor: moved to compatible with React 17
