+++ .eslintrc
@@ -2,7 +2,7 @@
   "extends": "eslint-config-airbnb",
   "env": {
     "browser": true,
-    "mocha": true,
+    "jest": true,
     "node": true
   },
   "parser": "babel-eslint",
