+++ .travis.yml
@@ -1,10 +1,10 @@
 language: node_js
 node_js:
-  - "8.5.0"
+  - "14"
 services:
   - xvfb
 before_script:
   - export DISPLAY=:99.0
 script:
-  - ./node_modules/karma/bin/karma start test/karma.conf.js --browsers Firefox --single-run --no-auto-watch --capture-timeout 60000
+  - npm run test
   - npm run lint
