+++ scripts/build.js
@@ -1,5 +1,4 @@
 var execSync = require('child_process').execSync;
-var readFileSync = require('fs').readFileSync;
 
 function exec(command) {
   execSync(command, { stdio: [0, 1, 2] });
